add separate chatgpt service and add logging service
